requirements.txt
Flask
flask-cors
Werkzeug
pymongo
python-dotenv
PyPDF2
python-docx
google-generativeai
chromadb
playwright

# Authentication and security
PyJWT>=2.8.0
bcrypt>=4.0.1

# Production-level requirements for profile analysis
selenium>=4.15.0
beautifulsoup4>=4.12.0
requests>=2.31.0
urllib3>=2.0.0
chromedriver-autoinstaller>=0.6.2
undetected-chromedriver>=3.5.4
fake-useragent>=1.4.0
aiohttp>=3.9.0
lxml>=4.9.0
python-dotenv>=1.0.0

# Optional AI dependencies
# langchain>=0.0.350
# openai>=1.3.0
